Only move code to Simple/GHC/Build*
